Simplify engine_initializer.py by moving setup of some options related code into proper rules
[ci skip-rust]

[ci skip-build-wheels]
